Clean up my garbage
Implement sys_native_char instead of sys_utf_char -> MUCH better code

+++ src/shl/string.cpp
@@ -860,7 +860,7 @@ static double _string_to_decimal(const_string_base<C> s, const_string_base<C> *n
             mant_size -= 1;
         }
 
-        fraction = frac1;
+        fraction = (double)frac1;
     }
 
     s = pexp;
